requirements.txt
asgiref==3.4.1
Babel==2.9.1
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2.5
django-autoslug==1.9.8
django-cities-light==3.8.1
django-js-asset==1.2.2
django-money==2.0
django-mptt==0.11.0
django-phone-field==1.8.1
django-phonenumber-field==5.2.0
django-rest-swagger==2.2.0
django-smart-selects==1.5.9
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
djoser==2.1.0
drf-yasg==1.20.0
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
MarkupSafe==2.0.1
oauthlib==3.1.1
openapi-codec==1.3.2
packaging==21.0
phonenumbers==8.12.26
Pillow==8.3.0
progressbar2==3.53.1
psycopg2-binary==2.9.1
py-moneyed==1.2
pycparser==2.20
PyJWT==2.1.0
pyparsing==2.4.7
python-utils==2.5.6
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
ruamel.yaml==0.17.10
ruamel.yaml.clib==0.2.6
simplejson==3.17.2
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
Unidecode==1.2.0
uritemplate==3.0.1
urllib3==1.26.6
